+++ include/entity/Miner.h
@@ -11,11 +11,11 @@ namespace Game3 {
 		public:
 			static Identifier ID() { return {"base", "entity/miner"}; }
 			constexpr static Index RADIUS = 50;
-			constexpr static float HARVESTING_TIME = 5.f;
-			constexpr static float SELLING_TIME = 5.f;
+			constexpr static double HARVESTING_TIME = 5.;
+			constexpr static double SELLING_TIME = 5.;
 
 			std::optional<Position> chosenResource;
-			float harvestingTime = 0.f;
+			double harvestingTime = 0.;
 
 			static std::shared_ptr<Miner> create(Game &);
 			static std::shared_ptr<Miner> create(Game &, RealmID overworld_realm, RealmID house_realm, const Position &house_position, const std::shared_ptr<Building> &keep_);
@@ -32,7 +32,7 @@ namespace Game3 {
 			friend class Entity;
 
 		protected:
-			float sellTime = 0.f;
+			double sellTime = 0.;
 
 			Miner();
 			Miner(RealmID overworld_realm, RealmID house_realm, const Position &house_position, const std::shared_ptr<Building> &keep_);
@@ -43,7 +43,7 @@ namespace Game3 {
 			void wakeUp();
 			void goToResource();
 			void startHarvesting();
-			void harvest(float delta);
+			void harvest(double delta);
 			void sellInventory();
 	};
