+++ include/item/EmptyFlask.h
@@ -6,6 +6,6 @@ namespace Game3 {
 	class EmptyFlask: public Item {
 		public:
 			using Item::Item;
-			bool use(Slot, ItemStack &, const Place &, Modifiers, std::pair<float, float>) override;
+			bool use(Slot, ItemStack &, const Place &, Modifiers, std::pair<double, double>) override;
 	};
 }
